refactor: TravelService -> TripService 이름 변경
